\name{eye_dat}
\docType{data}
\alias{eye_dat}
\title{Eye_Data}
\description{
eye-tracking data of an 11-month-old infant watching 15 real-world scene stimuli
}
\usage{data(eye_dat)}
\format{A dataframe with 4 variables (x- and y-coordinates of gazeposition in pixels, stimuli name and distance to the screen in mm)}
\source{S. P. Johnson}
\references{
  S. P. Johnson
}
\keyword{datasets}