\name{gazepath-package}
\alias{gazepat-package}
\docType{package}
\title{
Gazepath Transforms Eye-Tracking Data into Fixations and Saccades
}
\description{
Eye-tracking data must be transformed into fixations and saccades before it can be analyzed. This package provides a non-parametric speed-based approach to do this on a trial basis. The method is especially useful when there are large differences in data quality, as the thresholds are adjusted accordingly. The same pre-processing procedure can be applied to all participants, while accounting for individual differences in data quality.
}
\details{
\tabular{ll}{
Package: \tab gazepath\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-03\cr
License: \tab GPL-2\cr
}

}
\author{
Daan van Renswoude

Maintainer: Daan van Renswoude <D.R.vanRenswoude@uva.nl>
}
\references{
Mould, M. S., Foster, D. H., Amano, K., & Oakley, J. P. (2012). A simple nonparametric method for classifying eye fixations. Vision research, 57, 18-25.
}

\keyword{ package }

