\name{GUI}
\alias{GUI}
\title{
Graphical User Interface (GUI) for gazepath implemented in Shiny
}
\description{
This function is a wrapper around gazepath and provides a graphical user interface (GUI) in shiny to parse raw eye-tracking data into fixations and saccades. The GUI functions opens a browser that serves as GUI to parse and visualize eye-tracking data 
}
\usage{
GUI()
}

\arguments{
no arguments
}
\details{
no details
}
\value{
nothing is returned
}
\references{
Mould, M. S., Foster, D. H., Amano, K., & Oakley, J. P. (2012). A simple nonparametric method for classifying eye fixations. Vision research, 57, 18-25.
}
\author{
Daan van Renswoude & Ingmar Visser
}
\examples{
\dontrun{
GUI()
} 
}
\keyword{ fixations }
\keyword{ gazepath }
\keyword{ GUI }
