\name{screen}
\docType{data}
\alias{screen}
\title{Screen Dimensions}
\description{
Screen and stimuli dimensions in pixels and mm, belonging to the dataset 'eye_data'.
}
\usage{data(screen)}
\format{A dataframe with 4 variables (the height and width of stimuli in pixels and mm)}
\source{S. P. Johnson}
\references{
  S. P. Johnson
}
\keyword{datasets}