\name{gazepath-package}
\alias{gazepath}
\alias{gazepath}
\docType{package}
\title{
GazePath
}
\description{
Tool to transform raw eye-tracing data into fixations and saccades
}
\details{
\tabular{ll}{
Package: \tab gazepath\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-03\cr
License: \tab GPL-2\cr
}

}
\author{
Daan van Renswoude

Maintainer: Daan van Renswoude <D.R.vanRenswoude@uva.nl>
}
\references{
Mould et al. (2012)
}

\keyword{ package }

